renamed phpdoc extension to apidoc
