+++ extensions/yii/apidoc/components/BaseRenderer.php
@@ -1,17 +1,16 @@
 <?php
 /**
- * 
- * 
- * @author Carsten Brandt <mail@cebe.cc>
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\phpdoc\components;
-
+namespace yii\apidoc\components;
 
 use Yii;
 use yii\base\Component;
 use yii\console\Controller;
-use yii\phpdoc\models\Context;
+use yii\apidoc\models\Context;
 use yii\web\View;
 
 abstract class BaseRenderer extends Component
