+++ extensions/yii/apidoc/models/FunctionDoc.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\apidoc\models;
+
+class FunctionDoc extends BaseDoc
+{
+	public $signature;
+	public $input=array();
+	public $output;
+}
