+++ extensions/yii/apidoc/models/ParamDoc.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\apidoc\models;
+
+class ParamDoc
+{
+	public $name;
+	public $description;
+	public $type;
+	public $isOptional;
+	public $defaultValue;
+	public $isPassedByReference;
+}
