+++ extensions/yii/apidoc/models/PropertyDoc.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * 
- * 
- * @author Carsten Brandt <mail@cebe.cc>
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
  */
 
-namespace yii\phpdoc\models;
+namespace yii\apidoc\models;
 
 class PropertyDoc extends BaseDoc
 {
