--- extensions/yii/phpdoc/models/EventDoc.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace yii\phpdoc\models;
-
-class EventDoc extends BaseDoc
-{
-	public $isInherited;
-	public $definedBy;
-	public $trigger;
-}
