--- extensions/yii/phpdoc/models/FunctionDoc.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace yii\phpdoc\models;
-
-class FunctionDoc extends BaseDoc
-{
-	public $signature;
-	public $input=array();
-	public $output;
-}
