--- extensions/yii/phpdoc/models/ParamDoc.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * 
- * 
- * @author Carsten Brandt <mail@cebe.cc>
- */
-
-namespace yii\phpdoc\models;
-
-class ParamDoc
-{
-	public $name;
-	public $description;
-	public $type;
-	public $isOptional;
-	public $defaultValue;
-	public $isPassedByReference;
-}
